pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anki-vocab-automation"
version = "2.0.4"
description = "Automated vocabulary card creation for Anki using Collins Dictionary API"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Anki Vocabulary Automation Team"},
]
maintainers = [
    {name = "Anki Vocabulary Automation Team"},
]
keywords = ["anki", "vocabulary", "collins", "dictionary", "automation", "learning"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Education",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "beautifulsoup4>=4.9.0",
    "lxml>=4.6.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "flake8>=3.8.0",
    "black>=21.0.0",
    "isort>=5.0.0",
    "mypy>=0.900",
    "pre-commit>=2.0.0",
]
test = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "responses>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/Jason-Queen/anki-vocab-automation"
Documentation = "https://github.com/Jason-Queen/anki-vocab-automation#readme"
Repository = "https://github.com/Jason-Queen/anki-vocab-automation.git"
"Bug Tracker" = "https://github.com/Jason-Queen/anki-vocab-automation/issues"


[project.scripts]
anki-vocab = "anki_vocab_automation.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
anki_vocab_automation = ["templates/*.html"]

# Black code formatting
[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort import sorting
[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/conftest.py",
    "*/.venv/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# MyPy type checking
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "bs4.*",
    "lxml.*",
]
ignore_missing_imports = true 